UPSTREAM: drm: mali-dp: Clear CVAL when leaving config mode
It's possible for CVAL to get set whilst we are in config mode. If this
happens, afer we leave config mode the HW will latch whatever
configuration is in the registers at the next vsync. Most likely this
will be a partial configuration, as we'll be racing against the ongoing
atomic_commit.

To avoid this, clear CVAL before leaving config mode.

BUG=chromium:707244
TEST=Build and run

Change-Id: I169f766a19b6d9d62ea9121f70ee4c5321e8155f
Signed-off-by: Brian Starkey <brian.starkey@arm.com>
Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit e64053f05eb9)