net: danipc: Fix the missing interrupt issue
The HW clears all the sources including pending interrupts when SW
writes to the interrupt clearing register. The interrupt won't raise
again until more data are pushed into the FIFO.

To workaround this issue, before unmasking the interrupt, keep reading
the data from FIFO until it is empty after the interrupt is cleared.

Change-Id: I295cc548c04bb174026f3c35ce8018846ea5d0da
Acked-by: Jie Luo <jluo@qti.qualcomm.com>
Signed-off-by: Yasir Malik <ymalik@codeaurora.org>