ipc/sem: avoid idr tree lookup for interrupted semop
We can avoid the idr tree lookup (albeit possibly avoiding
idr_find_fast()) when being awoken in EINTR, as the semid will not change
in this context while blocked.  Use the sma pointer directly and take the
sem_lock, then re-check for RMID races.  We continue to re-check the
queue.status with the lock held such that we can detect situations where
we where are dealing with a spurious wakeup but another task that holds
the sem_lock updated the queue.status while we were spinning for it.  Once
we take the lock it obviously won't change again.

Being the only caller, get rid of sem_obtain_lock() altogether.

Link: http://lkml.kernel.org/r/1478708774-28826-3-git-send-email-dave@stgolabs.net
Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Cc: Manfred Spraul <manfred@colorfullife.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>