ovl: add infrastructure for intercepting file ops
Overlayfs needs to intercept a few file operations in order to properly
handle the following corner case:

 - file X is in overlayfs;

 - X resides on a lower (read-only) layer
	the lower file is L;

 - X is opened with O_RDONLY ->
	L is opened -> 'rofd';

 - X is opened with O_WRONLY ->
	this results in L being copied up to the upper layer file U
	U is opened -> 'rwfd';

 - write to 'rwfd' modifying U;

 - read from 'rofd' gets data from unmodified L;

While this is a rare occurrence, it has been known to cause problems.  To
prevent such inconsistencies, allow intercepting read operations and fix
them up in the unlikely case that it's needed.

This patch adds an ovl_fops structure that is going to contain the pieces
necessary for intercepting file operations:

 a) a new file_operations structure, based on the original but changing
    those operations that need to be intercepted;

 b) a pointer to the origin f_op.  Intercepted operations will normally
    just call the original, unless the file was copied up;

 c) a hash table entry to hook this up for reuse in subsequent opens.

The hash table is small (32 buckets) since there will only be a few
different file operations used (basically the number of different
filesystems used as lower layer of overlay).  The key to the has table is
the original fops pointer.

Insertion is done under a global mutex.  There will only be one insertion
event for each unique underlying file_operations structure, so this is
going to be rare.

The hash table is accessed using rcu, so this will add minimal overhead to
opens.  The override fops are removed only at module exit time, so we don't
even have to be worry about read-side rcu locking.

There are a few assumption this scheme makes about file operation
structures supplied by filesystems:

 1) the lifetime of the structures is equal to the lifetime of the
    filesystem module; ensure this by holding a ref to the
    sb->s_type->owner (normal filesystems don't fill in f_ops->owner).
    This means that once a filesystem has been used as lower layer of
    overlayfs its module cannot be removed until the overlay module itself
    is removed.  I don't believe this will be a problem.

 2) Filesystems should not make use of f_op in any way except possibly
    replacing it in ->open.  Overlayfs itself breaks this assumption, but
    recursion is handled by ->d_real so this is not an issue.  Add a
    ->magic field to the override fops structure to verify that the
    filesystem didn't mess with file->f_op.

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>