CHROMIUM: tpm: cr50: Start TPM on resume from suspend-to-idle
If using suspend-to-idle (S0ix) then the BIOS is not executed on resume,
but the TPM may go into a deep sleep state and lose context, so it needs
to be told to start up and restore that saved state.

However the TPM does not immediately go into a deep sleep state, so if
we wake up too early it may not have lost state and will return a code
indicating it is already initialized.  In order to prevent this from
generating a confusing error message in syslog the command is done without
a failure string, and the non-fatal return code is filtered out.

BUG=chrome-os-partner:59895
TEST=manual:
1) Test S3 suspend-to-ram to ensure it functions as usual with no changes.
2) Test S0ix suspend-to-idle with a long residency in suspend, and check
that the TPM context is restored when the system wakes.
3) Test S0ix suspend-to-idle with a short residency in suspend, and ensure
that there are no error messages in syslog after wake.

Change-Id: Ic8ba712b18afb7ea87b68fe7d35f07f6a508ece2
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/415535
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>