UPSTREAM: drm/bridge: analogix_dp: set the DPCD600 during disabling PSR
It looks like the BOE panel FW didn't ack the DPCD600 signal from the
host device, this will cause the panel to hang on the startup display.

The root cause is that we use the fast link mode when we enter and
exit PSR, the issue is gone if we switch from the fast link to main
link mode.

BUG=chromium:707244
TEST=Build and run

Change-Id: Id5e92b4a7a70e68a32c619c1f4cbf1e8b13696e5
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Signed-off-by: Archit Taneja <architt@codeaurora.org>
Link: http://patchwork.freedesktop.org/patch/msgid/1481338159-7189-1-git-send-email-wxt@rock-chips.com
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 2289b3e11f39)