z3fold: discourage use of pages that weren't compacted
If a z3fold page couldn't be compacted, we don't want it to be used for
next object allocation in the first place.  It makes more sense to add it
to the end of the relevant unbuddied list.  If that page gets compacted
later, it will be added to the beginning of the list then.

This simple idea gives 5-7% improvement in randrw fio tests and about 10%
improvement in fio sequential read/write.

Link: http://lkml.kernel.org/r/20161115170038.75e127739b66f850e50d7fc1@gmail.com
Signed-off-by: Vitaly Wool <vitalywool@gmail.com>
Cc: Dan Streetman <ddstreet@ieee.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>