BACKPORT: ktime: Cleanup ktime_set() usage
ktime_set(S,N) was required for the timespec storage type and is still
useful for situations where a Seconds and Nanoseconds part of a time value
needs to be converted. For anything where the Seconds argument is 0, this
is pointless and can be replaced with a simple assignment.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: Peter Zijlstra <peterz@infradead.org>

(cherry picked from drm-intel-next-queued 8b0e195314fabd58a331c4f7b6db75a1565535d7)

Forklift Note: Changes due to differing 4.9 usage of ktime