CHROMIUM: Include i2c_hid in the kernel for x86
This CL alters the x86 kernel config for ChromeOS devices to include
i2c_hid in the kernel.  Previously it was included as a module, but
then devices that use it are not availible early enough during the
boot process to manage their firmware updates.  With the driver compiled
into the kernel, the devices using it can be queried/updated before
the UI starts up.

BUG=chrome-os-partner:61563
TEST=compiled a kernel and deployed on an Electro.  On reboot the device
was present during the boot process

Change-Id: I2249bec552ef6401431b4e1d5afe3c8743067f16
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/425319
Reviewed-by: Benson Leung <bleung@chromium.org>