target: Remove several state tests from TMF code
Whether or not a session is being torn down does not affect whether
or not SCSI commands are in the task set. Hence remove the "tearing
down" checks from the TMF code. The CMD_T_FABRIC_STOP test is left
out because it overlaps with the CMD_T_COMPLETE and kref tests. The
TRANSPORT_ISTATE_PROCESSING check is left out because it is now
safe to wait for a command that is in that state.

Signed-off-by: Bart Van Assche <bart.vanassche@sandisk.com>
Reviewed-by: Hannes Reinecke <hare@suse.com>
Cc: Christoph Hellwig <hch@lst.de>
Cc: Andy Grover <agrover@redhat.com>
Cc: David Disseldorp <ddiss@suse.de>