ti-st: re-work instance handling
Multiple instance support is broken as an id of 0 is hardcoded in
several places and the FM radio driver only supports a single instance.

Rework st_kim_ref function to get per instance pointer from struct
device. For now the struct device is ignored and we just return the
single instance.

Signed-off-by: Rob Herring <robh@kernel.org>